P_Block. Шифрование. Рабочая версия программы.

+++ S_Block.cpp
@@ -4,7 +4,7 @@ S_Block::S_Block()
 {
 }
 
-int S_Block::Substitution(bitset<4> input)
+unsigned long S_Block::Substitution(bitset<4> input)
 {
 	return this->substitution[input.to_ulong()];
 }
